[ARCH] Adding db queries for users to its own file + adding api routes
